moved files from api folder into routes
