+++ packages/backend/index.js
@@ -1,20 +1,8 @@
-const express = require("express");
-const app = express(); 
-
-app.use(express.json())
-app.use(express.urlencoded({ extended: true }));
-
-app.use('/api', require('./api'))
-
-
-
-app.use((err, req, res, next) => {
-  console.error(err)
-  console.error(err.stack)
-  res.status(err.status || 500).send(err.message || 'Internal Server Error')
-})
+'use strict'
 
 const port = process.env.PORT || 5000;
+const app = require('./server')
+
 app.listen(port, () => {
   console.log(`server is listening on port ${port}`)
 })
